re structure

+++ bot/helper/listeners/tasks_listener.py
@@ -543,6 +543,7 @@ class MirrorLeechListener:
             msg += BotTheme('M_CC', Tag=self.tag)
             message = msg
             
+            btns = ButtonMaker()
             # <Section : MIRROR LOGS>
             if config_dict['MIRROR_LOG_ID']:
                 m_btns = deepcopy(buttons)
@@ -562,26 +563,27 @@ class MirrorLeechListener:
             if self.isPM and self.isSuperGroup:
                 message += BotTheme('M_BOT_MSG')
             buttons = extra_btns(buttons)[0]
-            if config_dict['SAFE_MODE']:
+            btns = extra_btns(btns)[0]
+            if self.isPM:
                 if self.isSuperGroup:
-                    s_btn = deepcopy(buttons)
+                    s_btn = deepcopy(btns) if config_dict['MIRROR_LOG_ID'] else deepcopy(buttons)
                     if self.source_url and config_dict['SOURCE_LINK']:
                         buttons.ubutton(BotTheme('SOURCE_URL'), self.source_url)
+                        if not config_dict['SAFE_MODE']:
+                            s_btn.ubutton(BotTheme('SOURCE_URL'), self.source_url)
                     if self.botpmmsg:
                         await sendMessage(self.botpmmsg, message, buttons.build_menu(2), photo=self.random_pic)
-                    if config_dict['SAVE_MSG'] and self.isSuperGroup:
-                        s_btn.ibutton(BotTheme('SAVE_MSG'), 'save', 'footer')
-                    if self.botpmmsg:
+                        if config_dict['SAVE_MSG']:
+                            s_btn.ibutton(BotTheme('SAVE_MSG'), 'save', 'footer')
                         s_btn.ibutton(BotTheme('CHECK_PM'), f"wzmlx {user_id} botpm", 'header')
                         await sendMessage(self.message, message, s_btn.build_menu(2), photo=self.random_pic)
                 else:
+                    if self.source_url and config_dict['SOURCE_LINK']:
+                        buttons.ubutton(BotTheme('SOURCE_URL'), self.source_url)
                     await sendMessage(self.message, message, buttons.build_menu(2), photo=self.random_pic)
             else:
-                if self.source_url and config_dict['SOURCE_LINK']:
+                if self.source_url and config_dict['SOURCE_LINK'] and ((not config_dict['SAFE_MODE'] and self.isSuperGroup) or (config_dict['SAFE_MODE'] and not self.isSuperGroup)):
                     buttons.ubutton(BotTheme('SOURCE_URL'), self.source_url)
-                if self.isPM and self.botpmmsg and self.isSuperGroup:
-                    await sendMessage(self.botpmmsg, msg, buttons.build_menu(2), self.random_pic)
-                    buttons.ibutton(BotTheme('CHECK_PM'), f"wzmlx {user_id} botpm", 'header')
                 if config_dict['SAVE_MSG'] and self.isSuperGroup:
                     buttons.ibutton(BotTheme('SAVE_MSG'), 'save', 'footer')
                 await sendMessage(self.message, message, buttons.build_menu(2), photo=self.random_pic)
@@ -597,7 +599,7 @@ class MirrorLeechListener:
                 await start_from_queued()
                 return
         
-        if not config_dict['DELETE_LINKS'] or config_dict['CLEAN_LOG_MSG']:
+        if self.botpmmsg and not config_dict['DELETE_LINKS'] or config_dict['CLEAN_LOG_MSG']:
             await deleteMessage(self.botpmmsg)
         
         await clean_download(self.dir)
